+++ app/app_haptics_engine_balanced_waveform.go
@@ -164,7 +164,7 @@ func (a *App) GenerateBalancedWaveform(rpm float64, engineRoughness float64, eng
 
 		// Add engine roughness similar to original implementation
 		if engineRoughness > 0.01 {
-			roughnessPhase := float64(a.state.current.sequenceNumber+uint32(index)) * 0.001
+			roughnessPhase := (float64(a.state.current.sequenceNumber) + float64(index)) * 0.001
 			roughnessContribution := math.Sin(roughnessPhase) * engineRoughness * 0.1
 			waveformValue += roughnessContribution
 		}
