+++ app/app_haptics_engine_torque_waveform.go
@@ -146,7 +146,7 @@ func (a *App) GenerateTorqueCurveWaveform(rpm float64, engineRoughness float64, 
 		combinedTorque := primaryComponent + secondaryComponent + tertiaryComponent
 
 		// Add engine-specific roughness modulation
-		roughnessPhase := float64(a.state.current.sequenceNumber+uint32(index)) * 0.0003
+		roughnessPhase := (float64(a.state.current.sequenceNumber) + float64(index)) * 0.0003
 		roughnessVariation := 1.0 + (engineRoughness * math.Sin(roughnessPhase) * 0.2)
 		combinedTorque *= roughnessVariation
